+++ src/main-extension-index.js
@@ -96,6 +96,9 @@ get('VERSION')
     .then((result)=>{
         console.log('setup is done');
         console.log(result);
+        
+        RCTF.setup_pane(RCTF.COLOR);
+        
     });
 });